# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CorpoPay API",
    "version": "0.1.0",
    "description": "Multi-tenant payment orchestration API."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "BillingEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "chargeId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ]
          },
          "currency": {
            "type": "string"
          },
          "attemptNumber": {
            "type": "number"
          },
          "billingPeriodStart": {
            "type": "string",
            "nullable": true
          },
          "billingPeriodEnd": {
            "type": "string",
            "nullable": true
          },
          "processedAt": {
            "type": "string",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "vpsTransactionId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "chargeId",
          "status",
          "amount",
          "currency",
          "attemptNumber",
          "billingPeriodStart",
          "billingPeriodEnd",
          "processedAt",
          "errorMessage",
          "vpsTransactionId"
        ]
      },
      "SubscriptionListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ]
          },
          "currency": {
            "type": "string"
          },
          "intervalType": {
            "type": "string"
          },
          "intervalValue": {
            "type": "number"
          },
          "nextBillingDate": {
            "type": "string",
            "nullable": true
          },
          "retryCount": {
            "type": "number"
          },
          "billingEventCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "customerId",
          "status",
          "amount",
          "currency",
          "intervalType",
          "intervalValue",
          "nextBillingDate",
          "retryCount",
          "billingEventCount",
          "createdAt"
        ]
      },
      "SubscriptionDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "initialPaymentIntentId": {
            "type": "string"
          },
          "paymentLinkId": {
            "type": "string",
            "nullable": true
          },
          "inngestRunId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ]
          },
          "currency": {
            "type": "string"
          },
          "intervalType": {
            "type": "string"
          },
          "intervalValue": {
            "type": "number"
          },
          "nextBillingDate": {
            "type": "string",
            "nullable": true
          },
          "currentPeriodStart": {
            "type": "string",
            "nullable": true
          },
          "currentPeriodEnd": {
            "type": "string",
            "nullable": true
          },
          "trialEndDate": {
            "type": "string",
            "nullable": true
          },
          "retryCount": {
            "type": "number"
          },
          "maxRetries": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "billingEvents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillingEvent"
            }
          }
        },
        "required": [
          "id",
          "tenantId",
          "customerId",
          "initialPaymentIntentId",
          "paymentLinkId",
          "inngestRunId",
          "status",
          "amount",
          "currency",
          "intervalType",
          "intervalValue",
          "nextBillingDate",
          "currentPeriodStart",
          "currentPeriodEnd",
          "trialEndDate",
          "retryCount",
          "maxRetries",
          "createdAt",
          "updatedAt",
          "billingEvents"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/subscriptions": {
      "get": {
        "operationId": "listSubscriptions",
        "summary": "List subscriptions",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "customerId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SubscriptionListItem"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "total",
                    "page",
                    "limit"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/subscriptions/{id}": {
      "get": {
        "operationId": "getSubscription",
        "summary": "Get subscription detail",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionDetail"
                }
              }
            }
          },
          "404": {
            "description": "Subscription not found"
          }
        }
      },
      "delete": {
        "operationId": "cancelSubscription",
        "summary": "Cancel a subscription",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "status"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Already cancelled"
          },
          "404": {
            "description": "Subscription not found"
          }
        }
      }
    },
    "/subscriptions/{id}/pause": {
      "post": {
        "operationId": "pauseSubscription",
        "summary": "Pause a subscription",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "status"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid state"
          },
          "404": {
            "description": "Subscription not found"
          }
        }
      }
    },
    "/subscriptions/{id}/resume": {
      "post": {
        "operationId": "resumeSubscription",
        "summary": "Resume a paused subscription",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "status"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid state"
          },
          "404": {
            "description": "Subscription not found"
          }
        }
      }
    },
    "/subscriptions/{id}/events": {
      "get": {
        "operationId": "listSubscriptionBillingEvents",
        "summary": "List billing events for a subscription",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BillingEvent"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "total",
                    "page",
                    "limit"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Subscription not found"
          }
        }
      }
    },
...
  }
}